requirements.txt
absl-py==2.1.0
accelerate==1.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
bitarray==3.2.0
bitsandbytes==0.43.3
blosc2==2.5.1
cachetools==5.5.0
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.8
-e git+ssh://git@github.com/mahmoodlab/CONCH.git@9a9a831cfb542e1c8e2e84017035be7de34541b8#egg=conch
contourpy==1.3.0
cycler==0.12.1
dashscope==1.20.12
docker-pycreds==0.4.0
ecos==2.0.14
einops==0.8.0
einops-exts==0.0.4
et-xmlfile==1.1.0
eval_type_backport==0.2.2
fast-pytorch-kmeans==0.2.0.1
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.2.3
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.1
h5py==3.11.0
huggingface-hub==0.27.0
idna==3.7
imageio==2.37.0
importlib_resources==6.4.4
inquirerpy==0.3.4
ipdb==0.13.13
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.14.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
ndindex==1.9.2
networkx==3.2.1
numexpr==2.10.2
numpy==1.25.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
nystrom-attention==0.0.12
open_clip_torch==2.29.0
opencv-python==4.10.0.84
openpyxl==3.1.5
openslide-python==1.3.1
osqp==1.0.4
pandas==2.0.3
peft==0.14.0
pfzy==0.3.4
pillow==10.4.0
propcache==0.2.0
protobuf==5.28.2
py-cpuinfo==9.0.0
pycparser==2.23
pydantic==2.10.4
pydantic_core==2.27.2
pynvml==11.5.3
pyparsing==3.1.4
pytz==2024.1
pyvips==3.0.0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
retrying==1.3.4
rich==13.9.4
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scikit-survival==0.23.1
scipy==1.13.1
sentry-sdk==2.19.2
setproctitle==1.3.4
smmap==5.0.1
sympy==1.13.2
tables==3.9.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==0.9.8
tokenizers==0.13.3
tomli==2.2.1
torch==2.4.0
torchmetrics==1.7.3
torchvision==0.19.0
tqdm==4.66.5
transformers==4.31.0
triton==3.0.0
tzdata==2024.1
urllib3==2.2.2
wandb==0.19.1
websocket-client==1.8.0
Werkzeug==3.0.4
yarl==1.16.0
